selection Sort test
