+++ algorithms/selectionSort/selectionSortSpec.md
@@ -1,0 +1,1 @@
+selectionSort Spec goes here!
